Moved new RPC methods to existing controllers
