+++ src/Stratis.Bitcoin.Features.Wallet/WalletRPCController.cs
@@ -1,6 +1,7 @@
-﻿using System;
-using System.Linq;
+﻿using System.Linq;
+using System.Threading.Tasks;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.Extensions.Logging;
 using NBitcoin;
 using Stratis.Bitcoin.Controllers;
 using Stratis.Bitcoin.Features.RPC;
@@ -11,15 +12,25 @@ namespace Stratis.Bitcoin.Features.Wallet
 {
     public class WalletRPCController : FeatureController
     {
-        public WalletRPCController(IServiceProvider serviceProvider, IWalletManager walletManager)
+        /// <summary>Instance logger.</summary>
+        private readonly ILogger logger;
+
+        /// <summary>Full node.</summary>
+        private readonly IFullNode fullNode;
+
+        /// <summary>Wallet broadcast manager.</summary>
+        private readonly IBroadcasterManager broadcasterManager;
+
+        /// <summary>Wallet manager.</summary>
+        private readonly IWalletManager walletManager;
+
+        public WalletRPCController(IWalletManager walletManager, IFullNode fullNode, IBroadcasterManager broadcasterManager, ILoggerFactory loggerFactory) : base(fullNode: fullNode)
         {
-            this.WalletManager = walletManager;
-            this.serviceProvider = serviceProvider;
+            this.walletManager = walletManager;
+            this.fullNode = fullNode;
+            this.broadcasterManager = broadcasterManager;
+            this.logger = loggerFactory.CreateLogger(this.GetType().FullName);
         }
-
-        internal IServiceProvider serviceProvider;
-
-        public IWalletManager WalletManager { get; set; }
 
         [ActionName("sendtoaddress")]
         [ActionDescription("Sends money to a bitcoin address.")]
@@ -29,13 +40,38 @@ namespace Stratis.Bitcoin.Features.Wallet
             return uint256.Zero;
         }
 
+        [ActionName("sendrawtransaction")]
+        [ActionDescription("Submits raw transaction (serialized, hex-encoded) to local node and network.")]
+        public async Task SendTransactionAsync(string hex)
+        {
+            this.logger.LogTrace("({0}:{1})", nameof(hex), hex);
+
+            Transaction transaction = this.fullNode.Network.CreateTransaction(hex);
+            await this.broadcasterManager.BroadcastTransactionAsync(transaction);
+
+            this.logger.LogTrace("(-)");
+        }
+
+        [ActionName("getnewaddress")]
+        [ActionDescription("Returns a new wallet address for receiving payments.")]
+        public string GetNewAddress()
+        {
+            this.logger.LogTrace("()");
+
+            HdAddress hdAddress = this.walletManager.GetUnusedAddress(this.GetAccount());
+            string base58Address = hdAddress.Address;
+
+            this.logger.LogTrace("(-):{0}", base58Address);
+            return base58Address;
+        }
+
         private WalletAccountReference GetAccount()
         {
             //TODO: Support multi wallet like core by mapping passed RPC credentials to a wallet/account
-            string w = this.WalletManager.GetWalletsNames().FirstOrDefault();
+            string w = this.walletManager.GetWalletsNames().FirstOrDefault();
             if (w == null)
                 throw new RPCServerException(RPCErrorCode.RPC_INVALID_REQUEST, "No wallet found");
-            HdAccount account = this.WalletManager.GetAccounts(w).FirstOrDefault();
+            HdAccount account = this.walletManager.GetAccounts(w).FirstOrDefault();
             return new WalletAccountReference(w, account.Name);
         }
     }
